correction to the gift-guide-cpt field definitions. the purpose of the second parameter is to define a text domain for translations if we ever wanted to translate the site into a different language.
in addition, changed _x to __. _x is used to translate as well, but it requires 3 parameters 1) default text 2) context of text to translate 3) text domain. In this case, only context is defined and text domain defaults to default. This would be fine if we wanted to translate separately based on context, but I can't see us ever needing to do that on makehub